[更新] 引导页&地图&更新放到appdelegate里 [更新] 消息列表footer更新 [更新] 优惠model 时间字段更新

+++ store/store/Service/Model/Store/TCPrivilege.h
@@ -23,8 +23,8 @@
 /** 可用时段 */
 @property (copy, nonatomic) NSArray *activityTime;
 /** 开始时间 */
-@property (assign, nonatomic) NSInteger startDate;
+@property (assign, nonatomic) int64_t startDate;
 /** 结束时间 */
-@property (assign, nonatomic) NSInteger endDate;
+@property (assign, nonatomic) int64_t endDate;
 
 @end
